:coffin: remove dead code, simplify and display data
